requirements.txt
python-docx>=1.1.2
Pillow>=10.0.0
